refactor: remove dead code and legacy paths
Drop helpers, attributes and branches nothing calls: runtime.get_max_gpu_memory, ITK._image_like, ome_zarr.update_konfai_attributes, vram.measure_transient_bytes, Dataset.get_statistics (a whole-volume read beside the streaming statistics), Attribute.is_info, CriterionsAttr.isTorchCriterion, the IMPACT Weights inner classes, the KONFAI_CONFIG_MODE=remove branch and the debug network blocks (Print, Write, Exit, Negative, GetShape, Const).

Remove the DDPM model, whose constructor has raised NotImplementedError since it was marked experimental; DiffusionGAN keeps the sinusoidal time embedding it took from it.

Drop the single-dash RESUME directory flags and the EMA avg_fn fallback for a torch without get_ema_multi_avg_fn (torch>=2.3 is required).
